MoacgScrapy/requirements.txt
aria2p==0.9.1
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
gevent==20.5.2
googletrans==3.0.0
greenlet==0.4.15
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.17
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.10
image==1.5.31
incremental==17.5.0
itemadapter==0.1.0
loguru==0.5.1
lxml==4.5.1
parsel==1.6.0
Pillow==7.1.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycurl==7.43.0.5
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-http-client==3.2.7
pytz==2020.1
queuelib==1.5.0
requests==2.23.0
rfc3986==1.4.0
Scrapy==2.2.0
selenium==3.141.0
sendgrid==6.3.1
service-identity==18.1.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
sqlparse==0.3.1
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.22.0
websocket-client==0.57.0
win32-setctime==1.0.1
wincertstore==0.2
youtube-dl==2020.6.16.1
zope.event==4.4
zope.interface==5.1.0
